requirements.txt
numpy>=1.24
Pillow>=9.0
matplotlib>=3.7

